\name{obtainPdbfrmPfam}
\alias{obtainPdbfrmPfam}
\title{obtainPdbfrmPfam}
\usage{
obtainPdbfrmPfam(searchItem = "xxx")
}
\description{
Get PDB id and its chain id from pfam accession number
}
\examples{
obtainPdbfrmPfam()
}
\references{
data source: "ftp://ftp.ebi.ac.uk/pub/databases/Pfam/mappings/pdb_pfam_mapping.txt"
}
